refactor: rename the component
